Bug fix
- improved Win popup animation
- fixed pack progress index show bug
- added energy bar in pause panel
- fixed timer save error

+++ Assets/App/Scripts/Architecture/Scene/Timer/TimerStateInfo.cs
@@ -5,7 +5,7 @@ namespace App.Scripts.Architecture.Scene.Timer
     public class TimerStateInfo
     {
         public double LastTotalSeconds;
-        public uint TimerSecondsTime;
+        public float TimerSecondsTime;
 
         public TimerStateInfo()
         {
